requirements.txt
aiohttp==3.6.1
aiomysql==0.0.19
asyncio-redis==0.15.1
dblpy==0.3.2
discord.py==1.2.3
imageio==2.4.1
imgurpython==1.1.7
jishaku==1.15.0.134
mysql-connector-python==8.0.13
nltk==3.3
numpy==1.16.1
oauthlib==3.0.2
objgraph==3.4.1
Pillow==6.2.0
psutil==5.5.0			# Get the dev version of python (sudo apt-get install python3.x-dev)
Pybooru==4.1.0
PyMySQL==0.9.2
PyNaCl==1.2.1
pyngrok==1.4.2
pynput==1.4.2
python-dateutil>=2.8.1
redis==2.10.6
requests==2.19.1
requests-oauthlib==1.2.0
urllib3==1.25.6
youtube-dl==2019.8.13
tabulate==0.8.5
imgurpython
nekos.py
tabulate